feat: changed arc http api to 'serverlessExpress', fixing gql codegen on dev-sls run
