+++ aws/src/api/graphql/categories.data.ts
