+++ aws/src/api/graphql/categories.defs.ts
