+++ aws/src/api/graphql/index.ts
@@ -1,9 +1,10 @@
-import { ApolloServer } from "apollo-server-lambda";
+import { ApolloServer } from "@apollo/server";
 
 import { typeDefs } from "./categories.defs";
 import { resolvers } from "./categories.resolvers";
 import { Categories } from "./categories.data";
 
+// Apply Apollo middleware to Express
 const rootValue = Categories;
 const server = new ApolloServer({ rootValue, typeDefs, resolvers });
-exports.handler = server.createHandler();
+export default server;
