+++ package.json
@@ -5,10 +5,10 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "sls": "npm run dev ---prefix aws",
+    "sls": "npm run dev --prefix aws",
     "dev-sls": "concurrently \"npm run sls\" \"npm run gql-codegen\" \"npm run dev\"",
     "netlify": "netlify dev",
-    "gql-codegen": "wait-on http-get://localhost:3000/api/graphql?query=%7B__typename%7D && graphql-codegen -c graphql/codegen.ts -w",
+    "gql-codegen": "wait-on http://localhost:3000/api/health && graphql-codegen -c graphql/codegen.ts -w",
     "build": "tsc && vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
